# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FleetFuel Portal API — target contract",
    "version": "0.1.0",
    "description": "Implementation handoff. No server exists yet. Validate against implementation in M06–M08. Business details: docs/04-BUSINESS-RULES.md and docs/05-API-CONTRACT.md."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local development after M00"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/vehicles": {
      "get": {
        "operationId": "listVehicles",
        "summary": "List tenant-scoped vehicles",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Vehicle"
                      }
                    },
                    "links": {
                      "$ref": "#/components/schemas/Links"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMeta"
                    }
                  },
                  "required": [
                    "data",
                    "links",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, revoked token or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role, ability, active-state or business-rule denial.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource absent or hidden by tenant scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed or no eligible product price.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds until retry."
              }
            }
          },
          "500": {
            "description": "Unexpected internal error; no stack trace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/per_page"
          },
          {
            "$ref": "#/components/parameters/company_id"
          }
        ]
      },
      "post": {
        "operationId": "createVehicle",
        "summary": "Create vehicle in allowed company",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Vehicle"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, revoked token or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role, ability, active-state or business-rule denial.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource absent or hidden by tenant scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed or no eligible product price.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds until retry."
              }
            }
          },
          "500": {
            "description": "Unexpected internal error; no stack trace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Laravel Sanctum personal access token; active role/ownership checks also required."
      }
    },
    "parameters": {
      "page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
      "per_page": {
        "name": "per_page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 25
        }
      },
...
      "company_id": {
        "name": "company_id",
        "in": "query",
        "description": "Admin only; all other callers must omit.",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              },
              "request_id": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "details",
              "request_id"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "Totals": {
        "type": "object",
        "properties": {
          "liters": {
            "type": "string",
            "pattern": "^[0-9]+(?:\\.[0-9]{1,2})?$",
            "description": "Plain decimal string, at most 2 places; response values use fixed scale."
          },
          "amount_lbp": {
            "type": "string",
            "pattern": "^[0-9]+(?:\\.[0-9]{1,2})?$",
            "description": "Plain decimal string, at most 2 places; response values use fixed scale."
          },
          "amount_usd": {
            "type": "string",
            "pattern": "^[0-9]+(?:\\.[0-9]{1,2})?$",
            "description": "Plain decimal string, at most 2 places; response values use fixed scale."
          }
        },
        "required": [
          "liters",
          "amount_lbp",
          "amount_usd"
        ],
        "additionalProperties": false
      },
      "PageMeta": {
        "type": "object",
        "properties": {
          "current_page": {
            "type": "integer",
            "minimum": 1
          },
          "per_page": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "last_page": {
            "type": "integer",
            "minimum": 1
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "totals": {
            "$ref": "#/components/schemas/Totals"
          }
        },
        "required": [
          "current_page",
          "per_page",
          "last_page",
          "total"
        ],
        "additionalProperties": false
      },
      "Links": {
        "type": "object",
        "properties": {
          "first": {
            "type": [
              "string",
              "null"
            ]
          },
          "last": {
            "type": [
              "string",
              "null"
            ]
          },
          "prev": {
            "type": [
              "string",
              "null"
            ]
          },
          "next": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "first",
          "last",
          "prev",
          "next"
        ],
        "additionalProperties": false
      },
...
      "VehicleRequest": {
        "type": "object",
        "properties": {
          "company_id": {
            "type": "integer",
            "minimum": 1,
            "description": "Required for admin; forbidden for manager."
          },
          "plate_no": {
            "type": "string",
            "minLength": 1,
            "maxLength": 30
          },
          "fuel_type": {
            "type": "string",
            "enum": [
              "petrol",
              "diesel"
            ]
          },
          "tank_capacity_l": {
            "type": "string",
            "pattern": "^[0-9]+(?:\\.[0-9]{1,2})?$",
            "description": "Plain decimal string, at most 2 places; response values use fixed scale."
          },
          "odometer_km": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          }
        },
        "required": [
          "plate_no",
          "fuel_type",
          "tank_capacity_l"
        ],
        "additionalProperties": false
      },
      "Vehicle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1
          },
          "company_id": {
            "type": "integer",
            "minimum": 1
          },
          "plate_no": {
            "type": "string"
          },
          "fuel_type": {
            "type": "string",
            "enum": [
              "petrol",
              "diesel"
            ]
          },
          "tank_capacity_l": {
            "type": "string",
            "pattern": "^[0-9]+(?:\\.[0-9]{1,2})?$",
            "description": "Plain decimal string, at most 2 places; response values use fixed scale."
          },
          "odometer_km": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "is_active": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "company_id",
          "plate_no",
          "fuel_type",
          "tank_capacity_l",
          "odometer_km",
          "is_active"
        ],
        "additionalProperties": false
      },
...
    }
  }
}